Added explicit symbol visibility markers to audio support files in ALmixer for public headers so they can be built as a dynamic library if needed.

+++ Isolated/ALmixer_RWops.c
@@ -8,7 +8,7 @@
 
 /* (Note this is different than stdio's seek. This returns ftell.)
  */
-static int stdio_seek(ALmixer_RWops* the_context, long offset, int whence)
+static long stdio_seek(ALmixer_RWops* the_context, long offset, int whence)
 {
 	if(0 == fseek(the_context->hidden.stdio.fp, offset, whence))
 	{
